Multi-line join

+++ src/code/string_array_functions.c
@@ -483,7 +483,7 @@ int stringIntoArray(char *str, char **array, int arraySize) {
  * @param arraySize
  * @return 1 for success, 0 for failure
  */
-int addStringIntoArray(char *string, char ***array, size_t arraySize) {
+int addStringIntoArray(char *string, char ***array, int arraySize) {
     if (string && array && arraySize >= 0) {
         char **tmp = realloc(*array, (arraySize + 1) * sizeof(char*));
         if (tmp) {
